// src/app/api/core/model/taskRecord.ts
/**
 * OpenAPI definition
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface TaskRecord { 
    taskId?: number;
    taskCreatorEmail?: string;
    taskTitle?: string;
    taskDescription?: string;
    taskStatus?: string;
    taskCritical?: string;
    taskStartTime?: Date;
    taskEndTime?: Date;
}